% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_by_cond.R
\name{abs_by_cond}
\alias{abs_by_cond}
\title{Compare Cluster abundances across conditions}
\usage{
abs_by_cond(
  seuratobject,
  condition = "Genotype",
  use_cols = NULL,
  font.total = 3,
  font.stack = 1.5,
  ctrl.group = "WT"
)
}
\arguments{
\item{seuratobject}{self explanantory}

\item{condition}{meta data variable you want to use, to compare the cluster abundances}

\item{use_cols}{named vector with the cluster identities as names and the value a custom color you want to use. If NULL standard color will be used.}

\item{font.total}{Font size of the total cell number (number above bar plot)}

\item{font.stack}{Font size of the cluster cell numbers in each stack}

\item{ctrl.group}{Name of the condition which should appear on the far left side of the plot}
}
\description{
This function creates a barplot, with stacks representing the single cell clusters and the x-axis the condition, which you specify by
parameter "condition". The y-axis shows the relativ abundance relative to the total cell number (number above bars) in the respective condition.
The numbers in the stacks show the cell number in the respective cell cluster
}
