% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_percent_zero.R
\name{cal_percent_zero}
\alias{cal_percent_zero}
\title{calculates percentage of zeros for each object in a list of seurat objects}
\usage{
cal_percent_zero(x)
}
\description{
calculates percentage of zeros for each object in a list of seurat objects
}
