% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.R
\name{doublets}
\alias{doublets}
\title{Wrapper for chris-mcginnis-ucsf/DoubletsFinder}
\usage{
doublets(x, dim = 1:30, sct = T, Ncpus = 10, pN = 0.25, exp_doub = 0.014)
}
\description{
Wrapper for chris-mcginnis-ucsf/DoubletsFinder
}
