% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_plots.R
\name{gen_plots}
\alias{gen_plots}
\title{wrapper for ggplot, in order to use ggplot for plotting multiple plots using lapply}
\usage{
gen_plots(col, data, list, xvalue, yvalue)
}
\description{
wrapper for ggplot, in order to use ggplot for plotting multiple plots using lapply
}
