% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_types.R
\name{gene_types}
\alias{gene_types}
\title{Gene Biotype distribution plot}
\usage{
gene_types(seurat_total, mart.use)
}
\arguments{
\item{seurat_total}{Seurat object}

\item{mart.use}{create Mart Object using UseMart()}
}
\description{
taken from scripts of Jonas Schrepping/DZNE
Shows a plot of the distribution of biotypes among the genes in a Seurat object
}
