% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{adapted from scripts of Jonas Schrepping/DZNE
GSEA function
need to execute after every restart, otherwise Error: external pointer is not valid
OrgDb = org.Mm.eg.db::org.Mm.eg.db}
\usage{
GSEA(
  object,
  condition = "Genotype",
  top = 50,
  GeneSets = "GO",
  Gontology = "BP",
  pCorrection = "bonferroni",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  min.pct = 0.1,
  logfc.threshold = 0.25,
  total = T,
  org = "mmu",
  present_genes = background,
  OrgDb,
  human_ensembl,
  mouse_ensembl
)
}
\description{
For proper functioning the package clusterProfiler and DOSE are required
}
