% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_SS2_kallisto.R
\name{import_SS2_kallisto}
\alias{import_SS2_kallisto}
\title{Expression Matrix from SS2 kallisto .h5 files}
\usage{
import_SS2_kallisto(dir, mart.use)
}
\arguments{
\item{dir}{Path to folder containing the transcript to genes list "tx2genes.csv" and a subdirectory called kallisto containing each .h5 file in a separate folder (folder name is
sample ID, e.g Plate and well number)}

\item{mart.use}{create Mart object using biomaRt::UseMart()}
}
\description{
Function imports single .h5 transcript abundance files and summarises them into a gene expression matrix (values normalised to length and sequencing depth).
The ensembl IDs are then transformed into gene names and a Seurat object is created. The values need still to be normalised and log transformed, according to the maintainers of
the tximport package (see https://support.bioconductor.org/p/132550/).
}
