% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_n_cluster.R
\docType{class}
\name{nbclust-class}
\alias{nbclust-class}
\alias{nbclust}
\title{determine the optimal number of clusters using Nbclust
find optimal #cluster using NbClust
select index to determine best number of clusters one of: "kl", "ch", "hartigan", "ccc", "scott", "marriot", "trcovw", "tracew", "friedman", "rubin", "cindex", "db", "silhouette", "duda", "pseudot2", "beale", "ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain", "gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw
if nothing is specified, all of them are used; if only.meta=T will skip the NbClust step and only create meta data with the nbclust clustering information from an existing nbclust class
ggplot necessities to be included in the function}
\description{
determine the optimal number of clusters using Nbclust
find optimal #cluster using NbClust
select index to determine best number of clusters one of: "kl", "ch", "hartigan", "ccc", "scott", "marriot", "trcovw", "tracew", "friedman", "rubin", "cindex", "db", "silhouette", "duda", "pseudot2", "beale", "ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain", "gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw
if nothing is specified, all of them are used; if only.meta=T will skip the NbClust step and only create meta data with the nbclust clustering information from an existing nbclust class
ggplot necessities to be included in the function
}
