% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_n_cluster.R
\name{optimal_n_cluster}
\alias{optimal_n_cluster}
\title{determine the optimal number of clusters using Nbclust
find optimal #cluster using NbClust}
\usage{
optimal_n_cluster(
  x,
  out = new("nbclust"),
  method = "kmeans",
  index = "all",
  min.nc = 1,
  max.nc = 15,
  only.meta = F
)
}
\arguments{
\item{x}{Seurat Object}

\item{method}{character, Clustering method, closest coming to Seurat clustering is kmeans}

\item{index}{Character vector, select indeces to determine best number of clusters one of: "kl", "ch", "hartigan", "ccc", "scott", "marriot", "trcovw", "tracew", "friedman", "rubin", "cindex", "db", "silhouette", "duda", "pseudot2", "beale", "ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain", "gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw
if nothing is specified, all of them are used}

\item{min.nc, max.nc}{integer, minimal and maximal number of clusters to calculate for}

\item{only.metaif}{will skip the NbClust step and only create meta data with the nbclust clustering information from an existing nbclust class}
}
\description{
determine the optimal number of clusters using Nbclust
find optimal #cluster using NbClust
}
