% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ambient.R
\name{remove_ambient}
\alias{remove_ambient}
\title{Ambient Genes Correction}
\usage{
remove_ambient(x, x_raw, cluster = Idents(x), meta.cols = c(1, 4:7))
}
\arguments{
\item{x}{processed SeuratObject}

\item{x_raw}{SeuratObject containing bad quality cells (before QC processing)}

\item{cluster}{Cluster identity for each cell (standard will take active Idents from SeuratObject x)}

\item{meta.cols}{columns from metadata to keep (see x[[]])... do not keep nCount_RNA and nFeature_RNA... needs to be recalculated}
}
\description{
correct for ambient genes using SoupX (github.com/constantAmateur/SoupX)
}
