% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCVolcano.R
\name{SCVolcano}
\alias{SCVolcano}
\title{DEGs single conditions}
\usage{
SCVolcano(
  x,
  Condition,
  cond1,
  cond2,
  logfc = 0.1,
  min.pct = 0.1,
  allvs1 = T,
  test = "wilcox.test",
  only.pos = F
)
}
\arguments{
\item{x}{SeuratObject}

\item{Condition}{name of metadata column to test for DEG}

\item{cond1, cond2}{Two conditions which are needed to be compared to each other(only if allvs1 is FALSE)}

\item{logfc, min.pct}{Cutoffs for find DEG}

\item{allvs1}{should just marker genes for each cluster be found (see Seurat::FindAllMarkers())}

\item{test}{which stastitical test to use: wilcoxon Rank sum, DESeq2, MAST, ... (see Seurat::FindAllMarkers())}
}
\description{
Performs DEG testing for single clusters as input for the function VolcanoPlot, to plot DEG for cells of a certain metadata group or clusters.
}
