% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_flow.R
\name{seurat_flow}
\alias{seurat_flow}
\title{Standard pipeline for seurat processing ... res: resolution for clustering; dim: number of dimensions to use (see ElbowPlot);
sct:perform SCTransform; magic: Perform imputation by magic? ... not done yet;
regress: character vector for metadata columns,which should be regressed for (e.g remove batch effects);
alg: clustering algorithm 1:Louvain 2:multilevel Louvain 3:SLM 4:Leiden(need to be installed);
low.features:TRUE if you are clustering for low number of genes (e.g CITE or Flow cytometry);
umap: "uwot"=R umap "umap-learn"= python package(faster, but needs to be installed)}
\usage{
seurat_flow(
  x,
  res = 0.7,
  dim = 1:30,
  sct = T,
  magic = F,
  norm = F,
  do.pca = T,
  regress = NULL,
  alg = 1,
  low.features = F,
  umap = "uwot",
  red.space = "pca"
)
}
\description{
Standard pipeline for seurat processing ... res: resolution for clustering; dim: number of dimensions to use (see ElbowPlot);
sct:perform SCTransform; magic: Perform imputation by magic? ... not done yet;
regress: character vector for metadata columns,which should be regressed for (e.g remove batch effects);
alg: clustering algorithm 1:Louvain 2:multilevel Louvain 3:SLM 4:Leiden(need to be installed);
low.features:TRUE if you are clustering for low number of genes (e.g CITE or Flow cytometry);
umap: "uwot"=R umap "umap-learn"= python package(faster, but needs to be installed)
}
