% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIs_cell.R
\name{UMIs_cell}
\alias{UMIs_cell}
\title{Plot #UMIs/cell vs. ranked cells}
\usage{
UMIs_cell(seurat_total, cutoff = 10000)
}
\arguments{
\item{seurat_total}{Seurat Object}

\item{cutoff}{At which #UMI to draw the cutoff line... this does not subset the Seurat Object, you still need to do it manually}
}
\description{
Show the number of UMIs per cells vs. the cells ranked by their UMI content. Linear correlation expected.
This plot helps to determine the #UMI cutoff, to exclude empty wells. The cutoff is set on the position, where the drop in UMIs is not directly proportional to the size of the cell.
taken from scripts of Jonas Schrepping/DZNE
}
