% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umis_genes_color.R
\name{umis_genes_color}
\alias{umis_genes_color}
\title{Scatterplot: #UMIs/cell vs. #genes/cell with color coded metadata}
\usage{
umis_genes_color(dataset, color = "percent.mito", dot.size = 0.5)
}
\arguments{
\item{dataset}{Seurat Object}

\item{color}{Name of the metadata column to use for colors in the plot. If FALSE no colorcoding will be performed.}

\item{dot.size}{The size of the single cells, represented by dots, in the plot}
}
\description{
Creates a Scatterplot, where the x axis represents the number of genes per cell and the y axis the number of UMIs per cell. The color can be set accordingly.
usually percentage of mitochondrial genes per cells is set as color, so that low quality cells (dying or damaged cells) can be identified by low number of UMIs,genes and high percentage
of mitochondrial genes.
}
