% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vln_pvalue.R
\name{Vln_pvalue}
\alias{Vln_pvalue}
\title{Signifinance Bars for Seurat's Violin Plots}
\usage{
Vln_pvalue(SeuratObject, gene, Condition, DEGs)
}
\arguments{
\item{SeuratObject}{on object of the Seurat class}

\item{gene}{Gene for which the significance test was performed}

\item{Condition}{name of metadata column to test for DEG}

\item{DEGs}{List of Differentially Expressed genes with their respective p values (output from FindMarkers)}
}
\description{
Takes adjusted p value from the output of FindMarkers and adds p values into the Violin plots (created by VlnPlot) of the respective gene.
}
