% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_stats.R
\name{align_stats}
\alias{align_stats}
\title{function taken from scripts of Jonas Schrepping/DZNE}
\usage{
align_stats(
  path,
  subset = F,
  exclude = F,
  legend.pos = "right",
  ymax = 2e+08,
  title = "Read Statistics"
)
}
\description{
finds QC statistics and summarises the information of multiple samples in one plot
}
