% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capwords.R
\name{capwords}
\alias{capwords}
\title{required for Seurat::AddModuleScore(), because often the gene names from other sources are not consisted in case.}
\usage{
capwords(s, strict = FALSE)
}
\description{
required for Seurat::AddModuleScore(), because often the gene names from other sources are not consisted in case.
}
