% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.R
\name{doublets}
\alias{doublets}
\title{Wrapper for chris-mcginnis-ucsf/DoubletsFinder}
\usage{
doublets(x, dim = 1:30, sct = T, Ncpus = 10, pN = 0.25, exp_doub = 0.014)
}
\arguments{
\item{x}{Seurat Object}

\item{dim}{integer sequence ... number of dimensions to use for Doublet finding}

\item{Ncpus}{integer, Number of cores to use for computation}

\item{pN}{numeric, metrics for calculation, see Doublet Finder vignette for Details}

\item{exp_doub}{numeric, expected percentage of doublets based on loaded#Cells/#Beads, formula: doublet rate= (#Cells/#Beads)^K*e^(-#Cells/#Beads)/K!, whereas K is 1, since we do not want to have more than one cell per bead}
}
\description{
Doublet Finder wrapper for usage on Seurat objects. You need to separately perform QC and clustering on your samples/Conditions before applying this function.
}
