% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_plot.R
\name{draw_plot}
\alias{draw_plot}
\title{Publication ready plots using ggplot2 ... any other layer can be added to this function by +, when executing
Parameter: y and xlab: axis titles; plot: which plot should be used( basically all geom_ ... possible );
legend: if want to remove use "none";do.facet= should facetting be performed, if yes type in facet the formula as string,
f.e "~Tissues" make separate plots for each tissue}
\usage{
draw_plot(
  obj,
  ylab = "y",
  xlab = "x",
  plot = ggplot2::geom_point(),
  legend = "right",
  do.facet = F,
  facet = NULL,
  x.fontsize = 15,
  y.fontsize = 15,
  legend.fontsize = 17,
  facets.fontsize = 18,
  facets.rotation = 45,
  x.rotation = 90,
  title.fontsize = 16,
  legend.key.size = 1.5,
  cluster_colors = NULL
)
}
\description{
Publication ready plots using ggplot2 ... any other layer can be added to this function by +, when executing
Parameter: y and xlab: axis titles; plot: which plot should be used( basically all geom_ ... possible );
legend: if want to remove use "none";do.facet= should facetting be performed, if yes type in facet the formula as string,
f.e "~Tissues" make separate plots for each tissue
}
