% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_types.R
\name{gene_types}
\alias{gene_types}
\title{taken from scripts of Jonas Schrepping/DZNE
Shows a plot of the distribution of biotypes among the genes in a Seurat object}
\usage{
gene_types(seurat_total)
}
\description{
taken from scripts of Jonas Schrepping/DZNE
Shows a plot of the distribution of biotypes among the genes in a Seurat object
}
