% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{Gene Set Enrichment Analysis of scRNAseq data}
\usage{
GSEA(
  object,
  condition = "Genotype",
  top = 50,
  GeneSets = "GO",
  Gontology = "BP",
  pCorrection = "bonferroni",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  min.pct = 0.1,
  logfc.threshold = 0.25,
  total = T,
  org = "mmu",
  present_genes = background,
  OrgDb,
  human_ensembl,
  mouse_ensembl,
  up = T
)
}
\arguments{
\item{object}{Seurat Object}

\item{condition}{name of column in metadata to use}

\item{top}{Number of genes from FindAllMarkers to use for GSEA}

\item{GeneSets}{which database to search (possible: GO,DO,KEGG,(from MSigDb:) Hallmark,cannonicalPathways,ImmunoSignatures,(Transcription Factor)Motifs)}

\item{Gontology}{more information in clusterprofiler vignette}

\item{pCorrection}{Choose the p-value adjustment method}

\item{pvalueCutoff, qvalueCutoff}{set the unadj. or adj. p-value (depending on correction method) and q-value cutoff (FDR corrected)}

\item{min.pct, logfc.threshold}{parameters for FindAllMarkers (see manual for further information)}

\item{total}{If TRUE, will perform bulk GSEA, not considering single seurat cluster (useful for total comparison between samples, or just for marker genes of the clusters)}

\item{org}{Specifying the species: mmu: mouse hsa: human. More information in clusterprofiler vignette}

\item{OrgDb}{mus musculus database need to be loaded prior to use}

\item{human_ensembl, mouse_ensembl}{use useMart("ensembl",dataset="...") to load the human and mouse databases}

\item{up}{should the upregulated or downregulated genes be used?}

\item{present_genes=background, #}{character vector, Background genes used for statistical testing: (universe in enricher() function), gene names need first to be translated into ensemblID using biomaRt}
}
\description{
Performs a GSEA for clusters, Conditions and Conditions between clusters. The GO,KEGG,DO, MSigDb databases are used.
For DO and MSigDB mouse genes are transformed to human homologues.
OrgDb = org.Mm.eg.db::org.Mm.eg.db need to execute after every restart, otherwise Error: external pointer is not valid.
Furthermore the Marts for translation from gene symbol to ensembl ID and finding homologues need to be loaded prior using this function (for usage see script in https://gitlab.dzne.de/frolova/spg15.git/scripts/Seurat_Flow_AF.Rmd)
the purpose of this function is to use it in foreach loop, in order to look for enrichment in all available databases in parallel (for usage see script in https://gitlab.dzne.de/frolova/spg15.git/scripts/Seurat_Flow_AF.Rmd).
For MSigDb GSEA To use the most up to date databases you might need to download the latest databases from MsigDb (https://www.gsea-msigdb.org/gsea/msigdb/).
The Databases used here were downloaded in July 2021.
Function adapted from scripts of Jonas Schrepping/DZNE.
}
