% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mtx.R
\name{read_mtx}
\alias{read_mtx}
\title{import multiple DropSeq-RNA samples using parallelisation
import expression matrices ... adds metadata from cell names (need to provide the names of the columns) and percentage
of zeros for each sample under dataset[[]][,zeros]... it is possible to convert ensmus ids to gene names}
\usage{
read_mtx(dir, meta_cols = c("Genotype", "Mouse", NA, "Pool"), ensmus = F)
}
\description{
import multiple DropSeq-RNA samples using parallelisation
import expression matrices ... adds metadata from cell names (need to provide the names of the columns) and percentage
of zeros for each sample under dataset[[]][,zeros]... it is possible to convert ensmus ids to gene names
}
