% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCVolcano.R
\name{SCVolcano}
\alias{SCVolcano}
\title{Volcano Plots for single cell clusters}
\usage{
SCVolcano(
  x,
  Condition,
  cond1,
  cond2,
  logfc = 0.1,
  min.pct = 0.1,
  do.col = F,
  lab.size = 7,
  title.size = 12,
  axis.font.size = 7,
  dot.size = 1,
  FC_cutoff = 0.25,
  p_cutoff = 0.05,
  legend = "none",
  xlab = NULL,
  ylab = NULL,
  axis.ticks = 0.3,
  axis.line = 0.3,
  allvs1 = T
)
}
\arguments{
\item{x}{SeuratObject}

\item{Condition}{name of metadata column to test for DEG}

\item{cond1, cond2}{Two conditions to plot in Volcano plot}

\item{logfc, min.pct}{Cutoffs for find DEG}

\item{do.col}{Should a color code be included?}

\item{lab.size, title.size, axis.font.size}{Change the fontsizes of the single graph elements}

\item{dot.size}{size of dots representing each gene in plot}

\item{legend}{Legend position see ggplot}

\item{xlab, ylab}{Genelabel you want to appear in the plot}

\item{axis.ticks, axis.line}{Thickness of ticks and lines}

\item{allvs1}{should just marker genes for each cluster be found (see Seurat::FindAllMarkers())}

\item{titles}{if used in loop, what titles should the single plots have? If you just look at clusters,this would be levels(x$cluster).}
}
\description{
wrapper for enhancedVolcano, made specifically for DE gene tables in Seurat 4.0 (careful Seurat <4.0 has "avg_logFC"
column instead of "avg_log2FC"). Performs DEG testing for single clusters and plots them separately into Volcano plots.
}
