% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_flow.R
\name{seurat_flow}
\alias{seurat_flow}
\title{Standard pipeline for seurat processing}
\usage{
seurat_flow(
  x,
  res = 0.7,
  dim = 1:30,
  sct = T,
  magic = F,
  norm = F,
  do.pca = T,
  regress = NULL,
  alg = 1,
  low.features = F,
  umap = "uwot",
  red.space = "pca",
  only.var = T
)
}
\arguments{
\item{res}{Float, resolution for clustering (see FindClusters)}

\item{dim}{Sequence of integers, Number of dimensions to use (see ElbowPlot)}

\item{sct}{Boolean Perform SCTransform?}

\item{magic}{Boolean, perform imputation by magic? ... not done yet}

\item{regress}{character vector for metadata columns,which should be regressed for (e.g remove batch effects)}

\item{alg}{clustering algorithm 1:Louvain 2:multilevel Louvain 3:SLM 4:Leiden(need to be installed)}

\item{low.features}{TRUE if you are clustering for low number of genes (e.g CITE or Flow cytometry)}

\item{umap}{"uwot"=R umap "umap-learn"= python package(faster, but needs to be installed)}

\item{red.space}{Character vector specifying the reduction space to use for FindNeigbors and RunUMAP}

\item{only.var}{Boolean, should only variable genes returned in scale.data of SCT Assay?}
}
\description{
Wrapper for Seurat tools. Allows to use different normalisation methods by changing one parameter. Furthermore, changes in dimensionality reduction can be adapted.
Magic imputation not developed yet.
}
