% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano_plot.R
\name{Volcano_plot}
\alias{Volcano_plot}
\title{wrapper for enhancedVolcano, made specifically for DE gene tables in Seurat 4.0 (careful Seurat <4.0 has "avg_logFC"
column instead of "avg_log2FC")}
\usage{
Volcano_plot(
  x,
  Condition,
  p_cutoff = 1e-05,
  FC_cutoff = 1,
  legend = "none",
  titles = levels(x$cluster),
  do.col = F
)
}
\description{
wrapper for enhancedVolcano, made specifically for DE gene tables in Seurat 4.0 (careful Seurat <4.0 has "avg_logFC"
column instead of "avg_log2FC")
}
