% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano_plot.R
\name{Volcano_plot}
\alias{Volcano_plot}
\title{Volcano Plots for DE genes in Seurat}
\usage{
Volcano_plot(
  x,
  Condition,
  p_cutoff = 1e-05,
  FC_cutoff = 1,
  legend = "none",
  titles = levels(x$cluster),
  do.col = F
)
}
\arguments{
\item{x}{data.frame generated by FindMarkers(), or in the same format.}

\item{Condition}{name of metadata column to use for color coding, used only if do.col=TRUE}

\item{p_cutoff, FC_cutoff}{Position in the plot where lines should be drawn, separating significant DE genes from non significant}

\item{legend}{Legend position see ggplot}

\item{titles}{if used in loop, what titles should the single plots have? If you just look at clusters,this would be levels(x$cluster).}

\item{do.col}{Should a color code be included?}
}
\description{
wrapper for enhancedVolcano, made specifically for DE gene tables in Seurat 4.0 (careful Seurat <4.0 has "avg_logFC"
column instead of "avg_log2FC")
}
